requirements.txt
astroid==3.3.8
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
dill==0.3.9
exceptiongroup==1.2.2
idna==3.10
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
packaging==24.2
platformdirs==4.3.6
pluggy==1.5.0
pylint==3.3.3
pytest==8.3.4
pytest-mock==3.14.0
python-dotenv==1.0.1
requests==2.32.3
soupsieve==2.6
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.12.2
urllib3==1.26.6
